application/backend/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
flask-mysql==1.5.2
flask-cors==3.0.10
pylint==2.17.0
pytest==7.2.2
black==23.1.0
virtualenv==20.21.0
python-dotenv==1.0.0
waitress==2.1.2